% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{Bdiag}
\alias{Bdiag}
\title{Create block diagonal matrix}
\usage{
Bdiag(A, k)
}
\arguments{
\item{A}{a numeric matrix with values in [0,1] forming each block.}

\item{k}{a numeric value indicating the number of blocks.}
}
\description{
Create block diagonal matrix
}
\examples{
Bdiag(matrix(1:4,2,2), 3)
}
