% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{invsqrt}
\alias{invsqrt}
\title{Create Block Diagonal matrix}
\usage{
invsqrt(A)
}
\arguments{
\item{A}{a numeric vectors with values in [0,1].}
}
\description{
Get an kernel function k(s,t) corresponds to an integral operator and a
basis system, return the corresponding matrix of the operator with respect
to given basis system.
}
