% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpsMat.r
\name{OpsMat}
\alias{OpsMat}
\title{The corresponding matrix of an integral operator}
\usage{
OpsMat(u, ker, basis)
}
\arguments{
\item{u}{a numeric vectors with values in [0,1].}

\item{ker}{a  bivariate numeric corresponds to an integral operator.}

\item{basis}{a functional basis object defining the basis.}
}
\value{
A numeric square matrix of the rank of basis system.
}
\description{
Get an kernel function k(s,t) corresponds to an integral operator and a
basis system, return the corresponding matrix of the operator with respect
to given basis system.
}
