% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{predict.sfar}
\alias{predict.sfar}
\title{Prediction of an SFAR model}
\usage{
\method{predict}{sfar}(object, h, ...)
}
\arguments{
\item{object}{a fit sfar object.}

\item{h}{number of steps ahead at which to predict.}

\item{...}{other.}
}
\description{
The h step ahead prediction for an SFAR(1) model.
}
