% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfar.r
\docType{package}
\name{Rsfar-package}
\alias{Rsfar-package}
\title{Rsfar: A Package for Seasonal Functional Autoregressive Models.}
\description{
The Rsfar package provides the collection of necessary
functions for simulating, estimating and forecasting seasonal functional
autoregressive time series of order one.
}
\details{
Functional autoregressive models are popular for functional time series analysis, but
the standard formulation fails to address seasonal behavior in functional time series data.
To overcome this shortcoming, we introduce seasonal functional autoregressive time series
models. For the model of order one, we provide estimation, prediction and simulation methods.
}
\references{
Atefeh Z., Hossein H., Maryam H., and R.J Hyndman (2021).
  Seasonal functional autoregressive models. Manuscript submitted for publication.
  \url{https://robjhyndman.com/publications/sfar/}
}
\seealso{
\code{\link{sfar}}, \code{\link{predict.sfar}},
}
