% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{Bdiag}
\alias{Bdiag}
\title{Create Block Diagonal matrix}
\usage{
Bdiag(A, k)
}
\arguments{
\item{A}{a numeric vectors with values in [0,1].}

\item{k}{a  bivariate}
}
\description{
Get an kernel function k(s,t) corresponds to an integral operator and a
basis system, return the corresponding matrix of the operator with respect
to given basis system.
}
