% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{Bdiag}
\alias{Bdiag}
\title{Create block diagonal matrix}
\usage{
Bdiag(A, k)
}
\arguments{
\item{A}{a numeric matrix forming each block.}

\item{k}{an integer value indicating the number of blocks.}
}
\value{
Return a block diagonal matrix from the matrix A.
}
\description{
Create block diagonal matrix
}
\examples{
Bdiag(matrix(1:4,2,2), 3)
}
