% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{invsqrt}
\alias{invsqrt}
\title{Return inverse square root of square positive-definite matrix.}
\usage{
invsqrt(A)
}
\arguments{
\item{A}{a numeric matrix}
}
\value{
inverse square root of square positive-definite matrix.
}
\description{
Return inverse square root of square positive-definite matrix.
}
\examples{
require(Rsfar)
X <- Bdiag(matrix(1:4,2,2), 3)
invsqrt(t(X) \%*\% X)
}
