% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfar.r, R/sfar_func.r
\docType{package}
\name{sfar}
\alias{sfar}
\title{sfar: A Package for Seasonal Functional Autoregressive Models.}
\usage{
sfar(
  X,
  seasonal,
  cpv = 0.85,
  kn = NULL,
  method = "MME",
  a = ncol(Coefs)^(-1/6)
)
}
\arguments{
\item{X}{a functional time series.}

\item{seasonal}{a positive integer variable specifying the seasonality paprameter.}

\item{cpv}{a numeric with values in [0,1] which determines the cumulative proportion variance explained by the first kn eigencomponents.}

\item{kn}{an integer variable specifying the number of eigencomponents.}

\item{method}{a character string giving the method of estimation. The following values are possible  :"MME" for Method of Moments and "ULSE" for Unconditional Least Square Estimation Method.}

\item{a}{a numeric with values in [0,1].}
}
\value{
A matrix of size p*p.
}
\description{
The sfar package provides the collection of necessary
functions  for simulating, estimating and forecasting seasonal functional
autoregressive (SFAR) time series of order one.

Estimate a seasonal functional autoregressive (SFAR) model of order 1 for a given functional time series.
}
\details{
Functional autoregressive models are popular for functional time series analysis, but
the standard formulation fails to address seasonal behavior in functional time series data.
To overcome this shortcoming, we introduce seasonal functional autoregressive time series
models. For the model of order one, we derive sufficient stationarity conditions and limiting behavior, and provide estimation and prediction methods. Moreover, we consider a
portmanteau test for testing the adequacy of this model and we derive its asymptotic distribution.
}
\examples{
# Generate Brownian motion noise
N <- 300 # the length of the series
n <- 200 # the sample rate that each function will be sampled
u <- seq(0, 1, length.out = n) # argvalues of the functions
d <- 45 # the number of bases
basis <- create.fourier.basis(c(0, 1), d) # the basis system
sigma <- 0.05 # the std of noise norm
Z0 <- matrix(rnorm(N * n, 0, sigma), nrow = n, nc = N)
Z0[, 1] <- 0
Z_mat <- apply(Z0, 2, cumsum) # N standard Brownian motion
Z <- smooth.basis(u, Z_mat, basis)$fd

# Simulate random SFAR(1) data
kr <- function(x, y) {
 (2 - (2 * x - 1)^2 - (2 * y - 1)^2) / 2
}
s <- 5 # the period number
X <- rsfar(kr, s, Z)
plot(X)

# SFAR(1) model parameter estimation:
Model1 <- sfar(X, seasonal = s, kn = 1)
}
\references{
Atefeh Z., Hossein H., Maryam H., and R.J Hyndman (2021).
  Seasonal functional autoregressive models. Manuscript submitted for publication.
}
\seealso{
\code{\link{sfar}}, \code{\link{predict.sfar}},
}
